Refactor type extenders and move parameters logic to its own folder
